+++ core/lib/constants.dart
@@ -130,6 +130,7 @@ const String keyData = "data";
 
 const String keyAndroidId = "isAndroidIdTrackingEnabled";
 const String keyAdId = "isAdIdTrackingEnabled";
+const String keyDeviceId = "isDeviceIdTrackingEnabled";
 
 // permission
 const String keyIsPermissionGranted = "isGranted";