+++ example/lib/main.dart
@@ -524,7 +524,19 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
                   _moengagePlugin.updatePushPermissionRequestCountAndroid(
                       pushPermissionCount);
                 },
-              )
+              ),
+              ListTile(
+                title: Text("Enable Device Id Tracking"),
+                onTap: () async {
+                  _moengagePlugin.enableDeviceIdTracking();
+                },
+              ),
+              ListTile(
+                title: Text("Disable Device Id Tracking"),
+                onTap: () async {
+                  _moengagePlugin.disableDeviceIdTracking();
+                },
+              ),
             ]).toList(),
           ),
         ),